requirements.txt
flask==3.0.3
gunicorn==22.0.0
Pillow==10.3.0
img2pdf==0.5.1
transformers==4.41.2
torch==2.3.0+cpu
torchvision==0.18.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
